Add per-frame interaction trace to TouchyPlayerController
The PlayerTick function only runs for the local player (i.e. only when
the PlayerController has a PlayerInput object), so it's the right place
for handling client-side player input.

A couple of properties in TouchyPlayerController configure our line
trace. These are currently set to sensible defaults for testing.

Each frame, we run a short line trace directly into the center of the
player's view, checking to see whether any primitive components in the
path of that line block the 'Interaction' trace channel. We also draw
some debug lines to visualize the results of the trace.