Add Interactible, an interface
An interface can't carry any data with it (so we can't expose properties
for configuration or maintain state), but it allows multiple disparate
classes to support the same set of functions. This gives us the same
functionality as AInteractibleActor, but without requiring that all our
interactible objects share a common base class.